fix: resolve all clippy warnings across nextgcore workspace
- Fix critical absurd_extreme_comparisons error in pool.rs (>= to ==)
- Replace inherent to_string() methods with Display impls (PlmnId, OgsSockaddr, Gtp2ApnIe, Ipv6Prefix, Dnn)
- Add workspace-level lint config for protocol-appropriate allows (type_complexity, too_many_arguments, dead_code, etc.)
- Add [lints] workspace = true inheritance to all 36 crates
- Auto-fix uninlined_format_args, redundant closures, unnecessary clones
- Fix manual slice copies, push-after-creation, field assignment patterns
- Fix unused variables, wildcard patterns, and missing imports
- Result: 0 clippy errors, 0 warnings, all tests pass

+++ src/bins/nextgcore-smfd/src/binding.rs
@@ -780,9 +780,9 @@ pub fn encode_traffic_flow_template(
         }
         
         let mut tft_pf = TftPacketFilter {
-            identifier: pf.identifier.saturating_sub(1) as u8,
+            identifier: pf.identifier.saturating_sub(1),
             direction: pf.direction as u8,
-            precedence: pf.precedence.saturating_sub(1) as u8,
+            precedence: pf.precedence.saturating_sub(1),
             content: PacketFilterContent::default(),
         };
